+++ config/medium/InversionCount/code/CGCode.js
@@ -1,0 +1,10 @@
+var inputs = readline().split(' ');
+const M = parseInt(inputs[0]);
+const A = parseInt(inputs[1]);
+const S = parseInt(inputs[2]);
+const N = parseInt(inputs[3]);
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('Inversion count');